Update IDA database for CTray
CTray methods highlighted in PINK exist on Vista, but not in XP. I did not
mark all just yet.

CTray methods highlighted in ORANGE are peculiar. They are marked this
colour to denote that they should be reanalysed in the future.

CTray methods highlighted in PURPLE are highlighted for analysis. They may
serve a very useful purpose.

I did not finish analysis of CTray yet, but most members are documented
correctly, and it is possible to reimplement most CTray functions just
fine with what is documented right now.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>