Fix control-flow bugs from commenting out code in tray.cpp
If you have an if conditional like:

if (condition)
    statement;
break;

and comment out the conditional `statement`, then you fall through to the
break. This is why such formatting of conditional statements is widely
considered to be bad design practice. If you replace all instances of
something throughout the whole codebase to comment it out, then you need
to pay special attention to the control flow to avoid messing it up.

In the above example, the code that the C compiler sees after preprocessing
would become:

if (conditional)
break;

which is equivalent to `if (conditional) break;`, but the formatting can be
misleading.

The mistake of using this style of conditional statement was prevalent in
Microsoft's original code, so it's not really our fault, and we have to
work around it.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>