Tray item work; fix notifications and fix GDI leak when changing icon
The GDI object leak when changing icons was particularly bad when the icon
changed many times rapidly, such as with Windows 8+ Advanced Task Manager
or Process Explorer. A hacky fix would have been to simply included the
`delete pni;` before the code Wiktor disabled in the past, but I decided to
fix notifications altogether.

Notifications were previously broken and caused an access violation in RPC
marshalling code due to some uninitialised properties. Wiktor must have
missed this in the past due to the NOTIFYITEM structure itself lacking many
of the properties that it has in Windows 10, but these properties were also
not initialised in the CNotificationItem object that is passed off to the
TNM_NOTIFY message handler of the tray notify window. At least, this is my
theory as to why the marshalling crashes, as adding code to initialise each
of these properties manually in CTrayItemManager::GetTrayItem fixed the
crash.

I have made additional preparations for bringing the tray in ExplorerEx to
parity with the canonical implementation in Windows 10 (well, I reference
ep_taskbar for convenience). This is not a particularly trivial feat, but
some things seem to make quite a bit of sense. For example, XP's tray can
only have a single notification CB registered to it, but any foreign
app can get the instance of the notification tray (via the undocumented yet
still-commonly-used ITrayNotify interface, an example is Chromium at
https://chromium.googlesource.com/chromium/src/+/8c47f6fe8fd5fa9fad817ec45bce63c4906a9b1c/chrome/browser/ui/views/status_icons/status_tray_state_changer_win.cc
) and register a callback, so it's worth considering support for multiple
callbacks being registered like 10 Explorer does.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>